+++ Example/AndesUI/Helper/Commons.swift
@@ -1,0 +1,33 @@
+//
+//  Commons.swift
+//  AndesUI-demoapp
+//
+//  Created by Daniel Esteban Beltran Beltran on 16/03/21.
+//  Copyright © 2021 MercadoLibre. All rights reserved.
+//
+
+import Foundation
+
+class Commons {
+    struct ComponentsTitle {
+        static let andesBadgeInTabBar = "AndesBadgeInTabBar"
+        static let andesTooltip = "AndesTooltip"
+        static let andesPogressIndicator = "AndesPogressIndicator"
+        static let andesBadge = "AndesBadge"
+        static let andesBottomSheet = "AndesBottomSheet"
+        static let andesButton = "AndesButton"
+        static let andesCard = "AndesCard"
+        static let andesCheckbox = "AndesCheckbox"
+        static let andesDatePicker = "AndesDatePicker"
+        static let andesCoachmark = "AndesCoachmark"
+        static let andesDropdown = "AndesDropdown"
+        static let andesList = "AndesList"
+        static let andesMessage = "AndesMessage"
+        static let andesRadioButton = "AndesRadioButton"
+        static let andesSnackbar = "AndesSnackbar"
+        static let andesTag = "AndesTag"
+        static let andesTextField = "AndesTextField"
+        static let andesThumbnail = "AndesThumbnail"
+        static let bottomSheetSwiftExampleViewController = "BottomSheetSwiftExampleViewController"
+    }
+}