+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 # 3.22.0
 ### ⚙️ Other
+- Home router refactored | Authors [jbrunoAtMeli] (https://github.com/jbrunoAtMeli)
 - Lib & testapp updated config: | Authors [jbrunoAtMeli] (https://github.com/jbrunoAtMeli)
     The following recommended warnings were enabled
         (In the library)