+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 # 3.23.0
 ### 🛠 Bug fixes
 - Fix color name and color in colorset | Authors [joalonspint](https://github.com/joalonsopint)
+### 🚀 Features
+- Added Analytics | Authors [dabeltran95](https://github.com/dabeltran95)
+
 
 # 3.22.0
 ### ⚙️ Other