+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-# Unpublished
+# 3.25.0
+### 🚀 Features
+- Change visibility public to private | Authors [jbrunoAtMeli](https://github.com/jbrunoAtMeli)
 ### 🛠 Bug fixes
 - Test app fix ProgressIndicatorRoute  andesBadgeExtensions| Authors [jbrunoAtMeli] (https://github.com/jbrunoAtMeli)
 - Proper horizontal align for radiobutton with no text label. Center radiobutton vertically with text | Author: [@vveltri](https://github.com/vveltri)